+++ tracker/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bitcoin_client;
 pub mod index;
+pub mod record;
 pub mod storage;
 
 pub use index::Index;
