+++ tracker/src/storage/memory.rs
@@ -1,0 +1,41 @@
+use crate::record::Record;
+use crate::storage::IndexStorage;
+use bitcoin::BlockHash;
+use std::cell::RefCell;
+use std::collections::HashMap;
+use std::convert::Infallible;
+
+#[derive(Debug)]
+pub struct MemoryIndexStorage(RefCell<HashMap<BlockHash, Vec<Record>>>);
+
+impl MemoryIndexStorage {
+    pub fn new() -> Self {
+        MemoryIndexStorage(RefCell::new(HashMap::new()))
+    }
+}
+
+impl IndexStorage for MemoryIndexStorage {
+    type Err = Infallible;
+
+    fn store_record(&self, record: Record) -> Result<(), Self::Err> {
+        let mut this = self.0.borrow_mut();
+        let vec = this.entry(record.bitcoin_block).or_default();
+        vec.push(record);
+        Ok(())
+    }
+
+    fn get_blocks_count(&self) -> Result<u64, Self::Err> {
+        Ok(self.0.borrow().len() as u64)
+    }
+
+    fn remove_with_block_hash(&self, hash: &BlockHash) -> Result<(), Self::Err> {
+        self.0.borrow_mut().remove(hash);
+        Ok(())
+    }
+
+    fn get_blocks_by_hash(&self, hash: &BlockHash) -> Result<Vec<Record>, Self::Err> {
+        let this = self.0.borrow();
+        let records = this.get(hash).map(Clone::clone).unwrap();
+        Ok(records)
+    }
+}
